+++ src/Groups/AllocationsGroup.php
@@ -4,7 +4,7 @@ namespace Spatie\FloatSdk\Groups;
 
 use Saloon\Http\BaseResource;
 use Saloon\Http\Response;
-use Spatie\FloatSdk\QueryParameters\GetAllocationsParameters;
+use Spatie\FloatSdk\QueryParameters\GetAllocationsParams;
 use Spatie\FloatSdk\Requests\GetAllocation;
 use Spatie\FloatSdk\Requests\GetAllocations;
 
@@ -15,7 +15,7 @@ class AllocationsGroup extends BaseResource
         return $this->connector->send(new GetAllocation($allocationId));
     }
 
-    public function all(?GetAllocationsParameters $parameters = null): Response
+    public function all(?GetAllocationsParams $parameters = null): Response
     {
         return $this->connector->send(new GetAllocations($parameters));
     }
