+++ src/Groups/ClientsGroup.php
@@ -4,7 +4,7 @@ namespace Spatie\FloatSdk\Groups;
 
 use Saloon\Http\BaseResource;
 use Saloon\Http\Response;
-use Spatie\FloatSdk\QueryParameters\GetClientsParameters;
+use Spatie\FloatSdk\QueryParameters\GetClientsParams;
 use Spatie\FloatSdk\Requests\GetClient;
 use Spatie\FloatSdk\Requests\GetClients;
 
@@ -15,7 +15,7 @@ class ClientsGroup extends BaseResource
         return $this->connector->send(new GetClient($clientId));
     }
 
-    public function all(?GetClientsParameters $parameters = null): Response
+    public function all(?GetClientsParams $parameters = null): Response
     {
         return $this->connector->send(new GetClients($parameters));
     }
