+++ src/Groups/ProjectsGroup.php
@@ -4,7 +4,7 @@ namespace Spatie\FloatSdk\Groups;
 
 use Saloon\Http\BaseResource;
 use Saloon\Http\Response;
-use Spatie\FloatSdk\QueryParameters\GetProjectsParameters;
+use Spatie\FloatSdk\QueryParameters\GetProjectsParams;
 use Spatie\FloatSdk\Requests\GetProject;
 use Spatie\FloatSdk\Requests\GetProjects;
 
@@ -15,7 +15,7 @@ class ProjectsGroup extends BaseResource
         return $this->connector->send(new GetProject($projectId));
     }
 
-    public function all(?GetProjectsParameters $parameters = null): Response
+    public function all(?GetProjectsParams $parameters = null): Response
     {
         return $this->connector->send(new GetProjects($parameters));
     }
