+++ src/Groups/UsersGroup.php
@@ -4,7 +4,7 @@ namespace Spatie\FloatSdk\Groups;
 
 use Saloon\Http\BaseResource;
 use Saloon\Http\Response;
-use Spatie\FloatSdk\QueryParameters\GetUsersParameters;
+use Spatie\FloatSdk\QueryParameters\GetUsersParams;
 use Spatie\FloatSdk\Requests\GetUser;
 use Spatie\FloatSdk\Requests\GetUsers;
 
@@ -15,7 +15,7 @@ class UsersGroup extends BaseResource
         return $this->connector->send(new GetUser($userId));
     }
 
-    public function all(?GetUsersParameters $parameters = null): Response
+    public function all(?GetUsersParams $parameters = null): Response
     {
         return $this->connector->send(new GetUsers($parameters));
     }
