+++ src/QueryParameters/GetAllocationsParams.php
@@ -4,7 +4,7 @@ namespace Spatie\FloatSdk\QueryParameters;
 
 use Spatie\FloatSdk\Concerns\HasPaginationAndSort;
 
-class GetProjectsParameters
+class GetAllocationsParams
 {
     use HasPaginationAndSort;
 
@@ -13,11 +13,16 @@ class GetProjectsParameters
      * @param  ?array<string>  $expand
      */
     public function __construct(
-        public ?bool $active = null,
-        public ?string $projectCode = null,
         public ?int $clientId = null,
-        public ?bool $nonBillable = null,
-        public ?string $tagName = null,
+        public ?int $projectId = null,
+        public ?int $phaseId = null,
+        public ?int $taskMetaId = null,
+        public ?int $peopleId = null,
+        public ?string $startDate = null,
+        public ?string $endDate = null,
+        public ?int $billable = null,
+        public ?int $status = null,
+        public ?int $repeatState = null,
         public ?string $modifiedSince = null,
         public ?array $fields = null,
         public ?array $expand = null,
@@ -36,17 +41,22 @@ class GetProjectsParameters
     public function toArray(): array
     {
         return array_filter([
-            'project_code' => $this->projectCode,
             'client_id' => $this->clientId,
-            'active' => is_null($this->active) ? null : (int) $this->active,
-            'nonBillable' => is_null($this->nonBillable) ? null : (int) $this->active,
-            'tag_name' => $this->tagName,
-            'page' => $this->page,
-            'per-page' => $this->perPage,
-            'sort' => $this->sort,
+            'project_id' => $this->projectId,
+            'phase_id' => $this->phaseId,
+            'task_meta_id' => $this->taskMetaId,
+            'people_id' => $this->peopleId,
+            'start_date' => $this->startDate,
+            'end_date' => $this->endDate,
+            'billable' => $this->billable,
+            'status' => $this->status,
+            'repeat_state' => $this->repeatState,
             'modified_since' => $this->modifiedSince,
             'fields' => $this->fields ? implode(',', $this->fields) : null,
             'expand' => $this->expand ? implode(',', $this->expand) : null,
+            'page' => $this->page,
+            'per-page' => $this->perPage,
+            'sort' => $this->sort,
         ], fn ($value) => $value !== null);
     }
 }
