+++ src/QueryParameters/GetProjectsParams.php
@@ -4,15 +4,21 @@ namespace Spatie\FloatSdk\QueryParameters;
 
 use Spatie\FloatSdk\Concerns\HasPaginationAndSort;
 
-class GetClientsParameters
+class GetProjectsParams
 {
     use HasPaginationAndSort;
 
     /**
-     * @param  ?array<int, string>  $fields
-     * @param  ?array<int, string>  $expand
+     * @param  ?array<string>  $fields
+     * @param  ?array<string>  $expand
      */
     public function __construct(
+        public ?bool $active = null,
+        public ?string $projectCode = null,
+        public ?int $clientId = null,
+        public ?bool $nonBillable = null,
+        public ?string $tagName = null,
+        public ?string $modifiedSince = null,
         public ?array $fields = null,
         public ?array $expand = null,
         ?int $page = null,
@@ -30,9 +36,15 @@ class GetClientsParameters
     public function toArray(): array
     {
         return array_filter([
+            'project_code' => $this->projectCode,
+            'client_id' => $this->clientId,
+            'active' => is_null($this->active) ? null : (int) $this->active,
+            'nonBillable' => is_null($this->nonBillable) ? null : (int) $this->active,
+            'tag_name' => $this->tagName,
             'page' => $this->page,
             'per-page' => $this->perPage,
             'sort' => $this->sort,
+            'modified_since' => $this->modifiedSince,
             'fields' => $this->fields ? implode(',', $this->fields) : null,
             'expand' => $this->expand ? implode(',', $this->expand) : null,
         ], fn ($value) => $value !== null);
