+++ src/Requests/GetProjects.php
@@ -5,7 +5,7 @@ namespace Spatie\FloatSdk\Requests;
 use Saloon\Enums\Method;
 use Saloon\Http\Request;
 use Saloon\Http\Response;
-use Spatie\FloatSdk\QueryParameters\GetProjectsParameters;
+use Spatie\FloatSdk\QueryParameters\GetProjectsParams;
 use Spatie\FloatSdk\Resources\ProjectResource;
 
 class GetProjects extends Request
@@ -13,7 +13,7 @@ class GetProjects extends Request
     protected Method $method = Method::GET;
 
     public function __construct(
-        protected ?GetProjectsParameters $parameters = null
+        protected ?GetProjectsParams $parameters = null
     ) {}
 
     protected function defaultQuery(): array
